Restructure
Restructure to split the client into two separate classes, one
responsible for initating rpc messages and another responsible for the
higher level kad client algorithms.

There seems to be a lot of duplication of code in the message passing
sections of the code. The multiplexing of external and internal messages
over one queue is a bit dissapointing (when compared with go having
multiple channel selects).